create utils dir to split utils into app and messaging utils
